Add files via upload
In this project we are going to make a Weather Station using a variety of sensors. The platform being used is an Arduino board. As we work through the project we will connect up the various sensors using I2C Bus protocol. In this process we create the software sketch that will run the weather station. We could also use a WiFi connection to send the data to the web server.